+++ dev/dev_history.R
@@ -97,7 +97,8 @@ usethis::use_build_ignore("_pkgdown.yml")
 # attachment::att_from_namespace()
 attachment::att_amend_desc(
   pkg_ignore = c("testthat", "dummypackage", "rstudioapi"),
-  extra.suggests = c("testthat", "pkgload", "rstudioapi") #, "MASS", "lattice")
+  extra.suggests = c("testthat", "pkgload", "rstudioapi")#,
+  # "MASS", "lattice", "Matrix")
 )
 # attachment::create_dependencies_file()