+++ tests/testthat/test-inflate.R
@@ -504,8 +504,7 @@ unlink(dummypackage, recursive = TRUE)
 alltemp <- tempfile("all.templates.inflate")
 dir.create(alltemp)
 
-for (pkgname in c("full", "teaching")) {
-  # TODO: "minimal" can inflate ? Empty dev_history does not work.
+for (pkgname in c("full", "teaching", "minimal")) {
   # No "additional" with create_fusen
   # {fusen} steps
   path_foosen <- file.path(alltemp, pkgname)
@@ -554,5 +553,73 @@ for (pkgname in c("full", "teaching")) {
 } # end of template loop
 # Delete dummy package
 unlink(alltemp, recursive = TRUE)
+
+# Tests empty chunks ----
+dummypackage <- tempfile("empty.chunks")
+dir.create(dummypackage)
+
+# {fusen} steps
+fill_description(pkg = dummypackage, fields = list(Title = "Dummy Package"))
+dev_file <- suppressMessages(add_flat_template(pkg = dummypackage, overwrite = TRUE, open = FALSE))
+flat_file <- dev_file[grepl("flat_", dev_file)]
+
+usethis::with_project(dummypackage, {
+  file.copy(
+    system.file("tests-templates/dev-template-empty.Rmd", package = "fusen"),
+    flat_file,
+    overwrite = TRUE
+  )
+  usethis::use_gpl_license()
+  # Add cars data
+  usethis::use_data(cars)
+
+  test_that("inflate() output no error", {
+    expect_error(
+      suppressMessages(
+        inflate(pkg = dummypackage, rmd = flat_file,
+                name = "Get started", check = FALSE)),
+      regexp = NA
+    )
+
+    # R files with chunk content - Name after title as function name is NA
+    pkgdoc <- file.path(dummypackage, "R", "my-pkg-doc.R")
+    expect_true(file.exists(pkgdoc))
+    pkgdoc_lines <- readLines(pkgdoc)
+    expect_equal(length(pkgdoc_lines), 10)
+    expect_equal(pkgdoc_lines[4], "\"_PACKAGE\"")
+
+    datadoc <- file.path(dummypackage, "R", "my-data-doc.R")
+    expect_true(file.exists(datadoc))
+    datadoc_lines <- readLines(datadoc)
+    expect_equal(length(datadoc_lines), 13)
+    expect_equal(datadoc_lines[13], "\"cars\"")
+
+    skip_if_not(interactive())
+    # Needs MASS, lattice, Matrix installed
+
+    checkdir <- tempfile("dircheck")
+    # Disable checking for future file timestamps
+    withr::with_envvar(
+      new = c(`_R_CHECK_SYSTEM_CLOCK_` = 0),
+      {
+        expect_error(
+          suppressMessages(
+            inflate(pkg = dummypackage, rmd = flat_file,
+                    name = "Get started", check = TRUE,
+                    check_dir = checkdir, quiet = TRUE,
+                    overwrite = TRUE)),
+          regexp = NA
+        )
+      })
+    # Should not be any errors with templates
+    check_lines <- readLines(file.path(checkdir, paste0(basename(dummypackage), ".Rcheck"), "00check.log"))
+    expect_equal(check_lines[length(check_lines)], "Status: OK")
+    unlink(checkdir, recursive = TRUE)
+  })
+})
+unlink(dummypackage, recursive = TRUE)
+
+# TODO - Add test for dev-template-two-fun-same-title.Rmd
+# TODO - Add test for dev-template-r6class.Rmd (dont check)
 
 # Do not create a second package with {fusen} in the same session, as it will mess up with `setwd()` and {usethis} needs these `setwd()`...